Move `Context` creation into `ContextStore`
This paves the way for collaborating within a context.
